Add Application class that represents the application.
The main() function is modified to create an instance of the Application
class that represents the application. Upon instantiation the Application
instance is initialised, for any required configuration, by reading
the config file. Once initialised, run() is called by main() to provide
a main application loop. The shutdown() method is added in order to
perform any cleanup that the application may need before the application ends.