Restructure application
Move the class definition of Stocktally (the main application)
to modules/application.py and refactor code so as to provide
for the basic structure of the application.

Reusable code for the application such as GetStartupMessage() and
GetShutdownMessage moved into the Stocktally class. Standard
console, user interface specific code, factored into Console class
defined in modules/console.py.

Stocktally.py creates the app object that represents the
application and the console that represents the standard console
UI.